# Delaunay graph computation