--- a/squidpy/gr/__init__.py
+++ b/squidpy/gr/__init__.py
@@ -1,5 +1,5 @@
 """The graph module."""
-from squidpy.gr._build import spatial_neighbors
+from squidpy.gr._build import spatial_neighbors, delaunay
 from squidpy.gr._nhood import nhood_enrichment, centrality_scores, interaction_matrix
 from squidpy.gr._ligrec import ligrec
 from squidpy.gr._ppatterns import ripley_k, co_occurrence, spatial_autocorr