### Types of changes
<!--- What types of changes are introduced? Put an `x` in all applicable boxes: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (change that would cause existing functionality to not work as before)
- [x] New feature (non-breaking change which adds functionality)

### Description
This PR adds the commutation of the Dealunay Triangulation graph form the ``scipy.spatial`` implementation. The graph is stored a adjacency matrix, similar to ``build_connectivities`` in ``_build.py`` 

### How has this been tested?
- ToDo -

### Closes
Related to #314