@giovp 2021-04-28
Hi @MxMstrmn ,

thanks a lot for the contribution! I had a look a bit at the current implementation and I think there is a faster solution without iterating across idxs for each simplex, you can try it out with this code

<details>
<summary>Details</summary>

```python
import scanpy as sc
import squidpy as sq
import scipy
import numpy as np
import itertools

adata = sq.datasets.visium_hne_adata()
from scipy.spatial import Delaunay, delaunay_plot_2d

points = adata.obsm["spatial"]
tri = Delaunay(points)

%%timeit
sq.gr.delaunay(
    adata,
)
# current 8.16 s ± 351 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

adj_current = adata.obsp["spatial_connectivities"]
sc.pl.spatial(
    adata,
    neighbors_key="spatial_neighbors",
    img_key=None,
    edges_width=1,
    edges=True,
)

%%timeit

N = adata.shape[0]
col_lst = []
row_lst = []
for i in np.arange(N + 1):
    idx = np.argwhere(i == tri.simplices)[:, 0]
    idx_col = np.unique(np.setdiff1d(tri.simplices[idx.squeeze(), ...], i)).tolist()
    col_lst.append(idx_col)
    row_lst.append(np.repeat(i, len(idx_col)))

col_indices = list(itertools.chain(*col_lst))
row_indices = list(itertools.chain(*row_lst))

# new 272 ms ± 8.25 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

adj = scipy.sparse.csr_matrix(
    (np.ones(len(row_indices)), (row_indices, col_indices)), shape=(N, N)
)
adata.obsp["spatial_connectivities"] = adj
sc.pl.spatial(
    adata,
    neighbors_key="spatial_neighbors",
    img_key=None,
    edges_width=1,
    edges=True,
)

np.testing.assert_array_equal(
    adj.toarray(),
    adj_current.toarray(),
)
# they are same
```

</details>

I'm gonna go on and cleanup a bit but there are couple of points that I'd like to discuss:

### qhull params exposed for users?
in https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.Delaunay.html there is note a mentioning that:

> Unless you pass in the Qhull option “QJ”, Qhull does not guarantee that each input point appears as a vertex in the Delaunay triangulation. Omitted points are listed in the coplanar attribute.

this refers to the `qhull_options`. Have you experimented a bit with this? do you think it's relevant? I'm wondering if it should be exposed to users, but if so a more detail explanation would be needed (which scipy doesn't provide).

### calculate distances for neighbors
Do you think it would be useful to calculate distances for neighbors? right now, it only outputs a `connectivities` obj but maybe `distances` would also be useful? I'm thinking because this method would be most likely used in non-grid structures where the distance actually have a meaning.

thanks in advance for inputs/suggestions