@MxMstrmn 2021-04-28
Hi @giovp,

Thanks for taking up the PR! I am much in favour of making the code faster, the speed up looks good to me but have not yet checked in detail. For the initial commit I wanted the code to be readable and tried to stay close to the original implementation for the KNN connectivities. If the results are the same and checked, that is great, thanks! 

## QHULL
I have not yet experimented with ``qhull_options`` and I am not sure when point do not show up in the simplexes, might be worth checking the ``coplanar`` attribute then? For the moment, I think this is not an issue but should definitely be kept in mind.

## Distances 
I agree, at least being able to provide the same functionality as for KNN makes sense. Especially as the distances might vary across simplexes.