@giovp 2021-04-28
ok I think this is good to go. There is some work to be done in the way arguments are handled in order to generalize graph construction but can be done in another PR (will summarize in an issue).

@MxMstrmn I added the euclidean distances as well, indeed re QULL useful to keep in mind,although don't have use case now to try it out, also documentation in qhull package is a bit cryptic