@michalk8 2021-04-28
Few comments:
- the condition starting https://github.com/theislab/squidpy/blob/delauny/squidpy/gr/_build.py#L163 seems to be broken,
  should be e.g.
```python
if delaunay:
   ...
else:
  tree = NearestNeighbors(n_neighbors=n_neigh or 6, radius=radius or 1, metric="euclidean")
  tree.fit(coords)
  if radius is None:
    ...
```