@giovp 2021-04-28
> the condition starting https://github.com/theislab/squidpy/blob/delauny/squidpy/gr/_build.py#L163 seems to be broken,
should be e.g.

done

> is this fast? https://github.com/theislab/squidpy/blob/delauny/squidpy/gr/_build.py#L176
generally, I'm a bit distrustful when using numpy + sets (or was this done to preserve the order + uniqueness?)

I think so yes, essentially I need to remove from the idx array the current `i` index. The whole graph building function on ~3k obs takes:
`701 ms ± 39.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

> should delauney be also passed here? https://github.com/theislab/squidpy/blob/delauny/squidpy/gr/_build.py#L93

good point, although logic should be change (see disc in #325 )

> could you also include this so that docs are being built? #323 (files)

done